+++ quasar-utils/src/test/groovy/net/corda/plugins/QuasarPluginTest.groovy
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.io.TempDir
 
 import java.nio.file.Path
 
-import static org.assertj.core.api.Assertions.*
-import static org.gradle.testkit.runner.TaskOutcome.*
+import static org.assertj.core.api.Assertions.assertThat
+import static org.gradle.testkit.runner.TaskOutcome.SUCCESS
 
 class QuasarPluginTest {
     private static final String TEST_GRADLE_USER_HOME = System.getProperty("test.gradle.user.home", ".")
@@ -20,6 +20,7 @@ class QuasarPluginTest {
     @BeforeEach
     void setup() {
         Utilities.installResource(testProjectDir, "settings.gradle")
+        Utilities.installResource(testProjectDir, "src/test/java/BasicTest.java")
     }
 
     @Test
@@ -36,6 +37,10 @@ repositories {
 }
 
 apply plugin: 'net.corda.plugins.quasar-utils'
+
+dependencies {
+    testImplementation 'junit:junit:4.12'
+}
 
 jar {
     enabled = false
@@ -81,6 +86,10 @@ repositories {
 
 apply plugin: 'net.corda.plugins.quasar-utils'
 
+dependencies {
+    testImplementation 'junit:junit:4.12'
+}
+
 jar {
     enabled = false
 }
@@ -115,6 +124,10 @@ repositories {
 
 apply plugin: 'net.corda.plugins.quasar-utils'
 
+dependencies {
+    testImplementation 'junit:junit:4.12'
+}
+
 jar {
     enabled = false
 }
@@ -148,13 +161,19 @@ repositories {
 
 apply plugin: 'net.corda.plugins.quasar-utils'
 
+dependencies {
+    testImplementation 'junit:junit:4.12'
+}
+
 jar {
     enabled = false
 }
 
 test {
-    allJvmArgs.forEach {
-        println "TEST-JVM: \${it}"
+    doLast {
+        allJvmArgs.forEach {
+            println "TEST-JVM: \${it}"
+        }
     }
 }
 """
@@ -165,19 +184,63 @@ test {
         }
     }
 
+    @Test
+    void checkOverriddenExclusionsAreUsed() {
+        def output = runGradleFor """
+plugins {
+    id 'net.corda.plugins.quasar-utils' apply false
+}
+
+description 'Show quasar-core added to test JVM arguments'
+
+repositories {
+    mavenCentral()
+}
+
+apply plugin: 'net.corda.plugins.quasar-utils'
+
+dependencies {
+    testImplementation 'junit:junit:4.12'
+}
+
+quasar {
+    excludePackages = [
+        "co.paralleluniverse**",
+        "groovy**"
+    ]
+}
+
+jar {
+    enabled = false
+}
+
+test {
+    doLast {
+        allJvmArgs.forEach {
+            println "TEST-JVM: \${it}"
+        }
+    }
+}
+"""
+        assertThat(output).anyMatch {
+            it.startsWith("TEST-JVM: -javaagent:") && it.endsWith('=x(co.paralleluniverse**;groovy**)')
+        }
+    }
+
     private List<String> runGradleFor(String script) {
         def buildFile = testProjectDir.resolve("build.gradle")
         buildFile.text = script
         def result = GradleRunner.create()
             .withProjectDir(testProjectDir.toFile())
-            .withArguments("--info", "--stacktrace", "build", "-g", TEST_GRADLE_USER_HOME)
+            .withArguments("--info", "--stacktrace", "test", "-g", TEST_GRADLE_USER_HOME)
             .withPluginClasspath()
+            .withDebug(true)
             .build()
         println result.output
 
-        def build = result.task(":build")
+        def build = result.task(":test")
         assertThat(build).isNotNull()
-        assertThat(build.outcome).isEqualTo(UP_TO_DATE)
+        assertThat(build.outcome).isEqualTo(SUCCESS)
 
         return result.output.readLines()
     }